+++ dev/com.ibm.ws.wsat.recovery_fat.multi.3/.settings/bndtools.core.prefs
@@ -1,0 +1,2 @@
+compileErrorAction=build
+eclipse.preferences.version=1
