+++ dev/com.ibm.ws.wsat.recovery_fat.multi.3/build.gradle
@@ -1,0 +1,69 @@
+/*******************************************************************************
+ * Copyright (c) 2020,2021 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+
+// Define G:A:V coordinates of each dependency
+dependencies {
+  requiredLibs project(":com.ibm.ws.tx.embeddable"),
+               project(":com.ibm.ws.wsat.recovery_fat.single"),
+               project(":com.ibm.ws.wsat.recovery_fat.multi.1"),
+               project(':io.openliberty.org.apache.commons.logging')
+}
+
+task copyCommonFiles {
+  doLast {
+    /*
+     * Copy test application resources.
+     */
+    copy {
+      from project(':com.ibm.ws.wsat.recovery_fat.single').file('test-applications')
+      into new File('test-applications')
+    }
+
+    /*
+     * Copy common publish files.
+     */
+    copy {
+      from project(':com.ibm.ws.wsat.recovery_fat.single').file('publish')
+      into new File('publish')
+    }
+
+    copy {
+      from project(':com.ibm.ws.wsat.recovery_fat.multi.1').file('fat/src/com/ibm/ws/wsat/fat/tests')
+      into new File('fat/src/com/ibm/ws/wsat/fat/tests')
+    }
+  }
+}
+
+jar {
+  dependsOn copyCommonFiles
+}
+
+autoFVT.dependsOn copyCommonFiles
+
+clean.doLast {
+  if (file('test-applications').exists()){
+    file('test-applications').deleteDir()
+  }
+  if (file('test-bundles').exists()){
+    file('test-bundles').deleteDir()
+  }
+  if (file('publish').exists()){
+    file('publish').deleteDir()
+  }
+  if (file('build').exists()){
+    file('build').deleteDir()
+  }
+  if (file('fat/src/com/ibm/ws/wsat/fat/tests').exists()){
+    file('fat/src/com/ibm/ws/wsat/fat/tests').deleteDir()
+  }
+}
+
+addRequiredLibraries.dependsOn addJakartaTransformer
