+++ dev/com.ibm.ws.wsat.recovery_fat.multi.4/.project
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<projectDescription>
+	<name>com.ibm.ws.wsat.recovery_fat.multi.4</name>
+	<comment></comment>
+	<projects>
+	</projects>
+	<buildSpec>
+		<buildCommand>
+			<name>org.eclipse.jdt.core.javabuilder</name>
+			<arguments>
+			</arguments>
+		</buildCommand>
+		<buildCommand>
+			<name>bndtools.core.bndbuilder</name>
+			<arguments>
+			</arguments>
+		</buildCommand>
+	</buildSpec>
+	<natures>
+		<nature>org.eclipse.jdt.core.javanature</nature>
+		<nature>bndtools.core.bndnature</nature>
+	</natures>
+	<linkedResources>
+		<link>
+			<name>com.ibm.ws.wsat.recovery_fat.single-recoveryClient</name>
+			<type>2</type>
+			<locationURI>PARENT-1-PROJECT_LOC/com.ibm.ws.wsat.recovery_fat.single/test-applications/recoveryClient/src</locationURI>
+		</link>
+		<link>
+			<name>com.ibm.ws.wsat.recovery_fat.single-recoveryServer</name>
+			<type>2</type>
+			<locationURI>PARENT-1-PROJECT_LOC/com.ibm.ws.wsat.recovery_fat.single/test-applications/recoveryServer/src</locationURI>
+		</link>
+	</linkedResources>
+</projectDescription>
