+++ dev/com.ibm.ws.wsat.recovery_fat.multi.4/wsat.recovery_fat.bnd
@@ -1,0 +1,16 @@
+#*******************************************************************************
+# Copyright (c) 2020,2021 IBM Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License v1.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+#
+# Contributors:
+#     IBM Corporation - initial API and implementation
+#*******************************************************************************
+-include= ~../cnf/resources/bnd/bundle.props
+bVersion=1.0
+
+#This file defines the main fat bundle
+
+Bundle-SymbolicName: ${basename;${basedir}}
